Update documentation and type definition organization. Remove unused dependencies.
