+++ packages/apputils/src/widgettracker.ts
@@ -11,6 +11,8 @@ import { FocusTracker, Widget } from '@phosphor/widgets';
 
 /**
  * A tracker that tracks widgets.
+ *
+ * @typeparam T - The type of widget being tracked. Defaults to `Widget`.
  */
 export interface IWidgetTracker<T extends Widget = Widget> extends IDisposable {
   /**
@@ -100,6 +102,8 @@ export interface IWidgetTracker<T extends Widget = Widget> extends IDisposable {
 
 /**
  * A class that keeps track of widget instances on an Application shell.
+ *
+ * @typeparam T - The type of widget being tracked. Defaults to `Widget`.
  *
  * #### Notes
  * The API surface area of this concrete implementation is substantially larger
