+++ packages/coreutils/package.json
@@ -32,13 +32,11 @@
     "watch": "tsc -b --watch"
   },
   "dependencies": {
-    "@phosphor/algorithm": "^1.1.2",
     "@phosphor/commands": "^1.6.1",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/disposable": "^1.2.0",
     "@phosphor/properties": "^1.1.2",
     "@phosphor/signaling": "^1.2.2",
-    "@phosphor/widgets": "^1.7.1",
     "ajv": "^6.5.5",
     "json5": "^2.1.0",
     "minimist": "~1.2.0",
