+++ tests/test-coreutils/package.json
@@ -14,15 +14,12 @@
   "dependencies": {
     "@jupyterlab/coreutils": "^3.0.0-alpha.11",
     "@jupyterlab/testutils": "^1.0.0-alpha.11",
-    "@phosphor/algorithm": "^1.1.2",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/disposable": "^1.2.0",
     "@phosphor/signaling": "^1.2.2",
-    "@phosphor/widgets": "^1.7.1",
     "chai": "^4.2.0",
     "jest": "^24.7.1",
     "jest-junit": "^6.3.0",
-    "simulate-event": "~1.4.0",
     "ts-jest": "^24.0.2"
   },
   "devDependencies": {
